Create list.md
This is the list assignment. It's not impressive really. The markdown file of my cheatsheet showcased a lot more stuff. It had ordered and unordered lists. And I was just happy with the way that was structured. It would be cool if you evaluated that as my list assignment. Thanks!